Add files via upload
this example shows how to move two steppers at the same time using the millis function. It removes the use of delays. The millis function could be changed with the function micros to make it more precise. Using it this way, the uptime of the project is around 30 minutes unless Arduino is reset.