Add files via upload
Last changes including an improvement to the bubble sorting and now it properly recognizes the 'biggest' number of steps and now it correlates correctly which stepper motor is doing an X amount of steps. The parallel_motors was changed and a few lines of code plus a function was added for the claw.