Add files via upload
Tracking of steps for art 1, art 3, art 4 , art 5 (1Motor) left to right. Fixed a bug that was increasing the count of steps by 1 when moving all the motors in parallel (fn parallel_motors). There's as always... some entanglement with the cables. Every function now counts the steps made by each motor so now it's a matter of putting a bit more of code to set the limits and range of movement.