Add files via upload
Now all the articulations follow the same sense of rotation. Positive steps mean CW and Negative steps mean CCW. A manual mode has been created. It has limits that can be adjusted based on the movement range of the arm. The inputs to control the robotic arm manually have to be entered one by one in order to avoid damage to the mechanical structure.